requirements.txt
streamlit>=1.39.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
azure-storage-blob>=12.0.0
pypdf>=4.0.0
docx2txt>=0.8